backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-multipart
pillow
numpy
torch
transformers
huggingface-hub
langchain
sentence-transformers
pgvector
redis
python-dotenv
reportlab
weasyprint
